enhanced the api

+++ app/controllers/event.js
@@ -1,4 +1,6 @@
 const Event = require('../models/Event')
+const HANDLER = require('../utils/response-helper')
+const HttpStatus = require('http-status-codes')
 module.exports = {
   createEvent: async (req, res, next) => {
     const event = new Event(req.body)
@@ -6,80 +8,72 @@ module.exports = {
       await event.save()
       res.status(201).json({ event: event })
     } catch (error) {
-      res.status(400).json({ error: error.errors })
+      res.status(400).json({ error: error })
     }
   },
   updateEvent: async (req, res) => {
-    Event.findById(req.params.id, (err) => {
-      if (err) {
-        res.status(400).json({ error: 'Did not find an event!' })
-      } else {
-        Event.findByIdAndUpdate(req.params.id, req.body, function (err, updatedEvent) {
-          if (err) {
-            res.status(400).json({ error: 'Cannot Update the data' })
-          } else {
-            res.status(201).json({ updatedEvent: updatedEvent })
-          }
-        })
+    const { id } = req.params
+    const updates = Object.keys(req.body)
+    try {
+      const event = await Event.findById(id)
+      if (!event) {
+        return res.status(HttpStatus.BAD_REQUEST).json({ message: 'No post exists' })
       }
-    })
+      updates.forEach(update => {
+        event[update] = req.body[update]
+      })
+      await event.save()
+      res.status(HttpStatus.OK).json({ event: event })
+    } catch (error) {
+      HANDLER.handleError(res, error)
+    }
   },
   rsvp: async (req, res) => {
     const { yes, no, maybe } = req.body
-    Event.findById(req.params.id, function (error, data) {
-      if (error) {
-        res.status(400).json({ error: error })
-      }
-      if (data == null) {
+    const { id } = req.params
+    try {
+      const data = await Event.findById(id)
+      if (!data) {
         res.status(400).json({ error: 'No Event is available' })
+        return
       }
-
       if (data.rsvpMaybe.includes(req.user.id) ||
       data.rsvpNo.includes(req.user.id) ||
       data.rsvpYes.includes(req.user.id)) {
         res.status(201).json({ msg: 'You have already done the rsvp' })
         return
       }
+      const event = await Event.findByIdAndUpdate(id)
       if (yes) {
-        Event.findByIdAndUpdate(req.params.id, {
-          $push: {
-            rsvpYes: req.user.id
-          }
-        },
-        function (error, data) {
-          if (error) {
-            res.status(400).json({ error: error })
-          }
+        try {
+          event.rsvpYes.push(req.user.id)
+          await event.save()
           res.status(201).json({ rsvpData: data })
-        })
+        } catch (error) {
+          res.status(400).json({ error: error })
+        }
       }
       if (no) {
-        Event.findByIdAndUpdate(req.params.id, {
-          $push: {
-            rsvpNo: req.user.id
-          }
-        },
-        function (error, data) {
-          if (error) {
-            res.status(400).json({ error: error })
-          }
+        try {
+          event.rsvpNo.push(req.user.id)
+          await event.save()
           res.status(201).json({ rsvpData: data })
-        })
+        } catch (error) {
+          res.status(400).json({ error: error })
+        }
       }
       if (maybe) {
-        Event.findByIdAndUpdate(req.params.id, {
-          $push: {
-            rsvpMaybe: req.user.id
-          }
-        }, function (error, data) {
-          if (error) {
-            res.status(400).json({ error: error })
-          }
+        try {
+          event.rsvpMaybe.push(req.user.id)
+          await event.save()
           res.status(201).json({ rsvpData: data })
-        })
+        } catch (error) {
+          res.status(400).json({ error: error })
+        }
       }
+    } catch (error) {
+      HANDLER.handleError(res, error)
     }
-    )
   },
   GetEventById: async (req, res, next) => {
     const { id } = req.params
@@ -109,13 +103,14 @@ module.exports = {
   deleteEvent: async (req, res, next) => {
     const { id } = req.params
     try {
-      const event = await Event.findByIdAndRemove(id)
-      if (!event) {
-        return res.status(400).json({ error: 'No such Event exists!' })
+      const deleteEvent = await Event.findById(id)
+      if (!deleteEvent) {
+        return res.status(HttpStatus.NOT_FOUND).json({ message: 'No Event exists' })
       }
-      res.status(201).json({ DeletedEvent: event })
+      await Event.findByIdAndRemove(id)
+      res.status(HttpStatus.OK).json({ deleteEvent: deleteEvent, message: 'Deleted the event' })
     } catch (error) {
-      next(error)
+      HANDLER.handleError(res, error)
     }
   }
 }
